feat: Overlay图层鸿蒙化适配 (#24)

+++ harmony/maps/src/main/cpp/MapsJSIBinder.h
@@ -263,7 +263,7 @@ namespace rnoh {
             object.setProperty(rt, "bounds", "array");
             object.setProperty(rt, "bearing", "int");
             object.setProperty(rt, "tappable", "bool");
-            object.setProperty(rt, "opacity", "int");
+            object.setProperty(rt, "opacity", "float");
             return object;
         }
     };
